refactor(gui-v2): corrections and improvements
