+++ packages/nc-gui-v2/components/monaco/Editor.vue
