+++ packages/nc-gui-v2/pages/project/index/create.vue
@@ -5,7 +5,7 @@ import { useToast } from 'vue-toastification'
 import { nextTick, ref } from '#imports'
 import { navigateTo, useNuxtApp } from '#app'
 import { extractSdkResponseErrorMsg } from '~/utils/errorUtils'
-import { projectTitleValidator } from '~/utils/projectCreateUtils'
+import { projectTitleValidator } from '~/utils/validation'
 import MaterialSymbolsRocketLaunchOutline from '~icons/material-symbols/rocket-launch-outline'
 
 const name = ref('')
