+++ packages/nc-gui-v2/utils/projectCreateUtils.ts
@@ -175,23 +175,6 @@ export const getDefaultConnectionConfig = (client: ClientType): ProjectCreateFor
   }
 }
 
-export const projectTitleValidator = {
-  validator: (rule: any, value: any, callback: (errMsg?: string) => void) => {
-    if (value?.length > 50) {
-      callback('Project name exceeds 50 characters')
-    }
-    if (value[0] === ' ') {
-      callback('Project name cannot start with space')
-    }
-
-    callback()
-  },
-}
-export const fieldRequiredValidator = {
-  required: true,
-  message: 'Field is required',
-}
-
 export const sslUsage = ['No', 'Preferred', 'Required', 'Required-CA', 'Required-IDENTITY']
 
 export const generateUniqueName = () => {
